# tar/storage: Replace TeeReader with MultiWriter